Moved guid to it's own file
