+++ src/metal/guid-key.js
@@ -1,0 +1,1 @@
+export default 'shopify-buy-uuid';
