+++ src/models/reference-model.js
@@ -1,5 +1,5 @@
 import BaseModel from './base-model';
-import { GUID_KEY } from '../metal/set-guid-for';
+import GUID_KEY from '../metal/guid-key';
 
 const ReferenceModel = BaseModel.extend({
