+++ tests/unit/api/shop-client-test.js
@@ -3,7 +3,7 @@ import { step, resetStep } from 'shopify-buy/tests/helpers/assert-step';
 import ShopClient from 'shopify-buy/shop-client';
 import Config from 'shopify-buy/config';
 import CartModel from 'shopify-buy/models/cart-model';
-import { GUID_KEY } from 'shopify-buy/metal/set-guid-for';
+import GUID_KEY from 'shopify-buy/metal/guid-key';
 
 const configAttrs = {
   domain: 'buckets-o-stuff.myshopify.com',
